Refactor seed data into modular entity files
